+++ yike-20260707/alibabacloud_yike20260707/models/_get_video_translation_job_request.py
@@ -9,6 +9,8 @@ class GetVideoTranslationJobRequest(DaraModel):
         self,
         job_id: str = None,
     ):
+        # The task ID returned by `SubmitVideoTranslationJob`.
+        # 
         # This parameter is required.
         self.job_id = job_id